Create NewDiceGame
This is the old dice game, but the Javascript and HTML have been changed to allow 3 players to play.  Some elements, when I tried to change from document.getElementByID() to the JQuery $("#id name"); , for some reason the game would not work.  I'll see if I can figure it out tomorrow, but if you have any ideas, let me know.